Move validation tests out of the route test file
